opcodes.csv: add replacement for undoc opcodes on wdc
the original wdc model replaced all undocumented opcodes
of the 6502 with nops (1-byte,1-cycle) in the columns
$_3, $_7, $_B and $_F [0].
the other undocumented opcodes outside these four columns
were replaced either with new functions, single-cycle
nops if the previous undocumented opcode was *not* a nop,
or with multi-cycle nops with "backwards-compatible"
address modes *iff* the previous undocumented opcode
implemented was also a multi-cycle nop.
however if the previous undocumented opcode had an
optional page-cross penalty cycle, the address mode was
simplified to one that doesn't have the penalty, as is
the case e.g. with 0xdc, which changed from abx to abs[2].
i marked all the new nop opcodes with '%' prefix.
additionally the wdc model featured the STP and WAI
instructions that were removed from the rockwell/wdc
models, but later re-introduced in the 65816[1].
these are marked with '+' prefix for the time being.

[0]: http://forum.6502.org/viewtopic.php?t=5918#p72952
[1]: http://www.oxyron.de/html/opcodesc02.html
[2]: http://www.oxyron.de/html/opcodes02.html